Code for Parameterizing agent-based models using an online game
Niko Yiannakoulias, Michel Grignon, Tara Marshall,
Parameterizing agent-based models using an online game,
Computers, Environment and Urban Systems,
Volume 112,
2024,
102142,
ISSN 0198-9715,
https://doi.org/10.1016/j.compenvurbsys.2024.102142.
(https://www.sciencedirect.com/science/article/pii/S0198971524000711)
Abstract: Agent-based models (ABMs) of human systems are often parameterized using real-world data. For some ABMs this is not possible because the reality upon which the models are based does not exist or is not generalizable from one setting to another. In this paper we implement an online decision game to parameterize an agent-based model of pedestrian and cyclist route choice decisions in a neighbourhood. Our conceptual framework is to use an experimental game to log decision-making behaviour, summarize this behaviour into a decision model, and then transfer this model to an ABM. The product of this framework is an ABM with agents informed by human decision making made within the game, rather than the real world. The results of our analysis suggest that the decision model is consistent with some general theory about decision making, but the ABM illustrates some unique and contextually specific patterns of flow. ABMs parameterized with game data may be useful for forecasting the effects of change on urban transportation infrastructure.
Keywords: Research gaming; Agent-based models; Online data collection; Route choice